refactor: move controllers file
