+++ src/infra/http/nest/controllers/associate/delete.controller.spec.ts
