+++ src/infra/http/nest/controllers/associate/delete.controller.ts
