+++ src/infra/http/nest/controllers/associate/find-all.controller.ts
