+++ src/infra/http/nest/controllers/associate/find-by-id.controller.ts
