+++ src/infra/http/nest/controllers/payment/delete.controller.ts
