+++ src/infra/http/nest/controllers/payment/find-all.controller.ts
