+++ src/infra/http/nest/controllers/payment/find-by-id.controller.ts
