+++ src/infra/http/nest/modules/auth.module.ts
@@ -3,13 +3,13 @@ import { PassportModule } from '@nestjs/passport';
 import { JwtModule, JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
 import { Env } from '@/env';
-import { AuthController } from './auth.controller';
-import { UserController } from './user.controller';
+import { AuthController } from '../controllers/auth/auth.controller';
+import { UserController } from '../controllers/auth/user.controller';
 import { CreateUserService } from '@/domains/auth/services';
 import { UserRepositoryPrismaAdapter } from '@/infra/repositories/prisma/domains/auth/user.repository.adapter';
 import { PrismaService } from '@/infra/repositories/prisma/prisma.service';
 import { CreateSessionService } from '@/domains/auth/services/create-session.service';
-import { JwtAdapter } from './jwt.provider';
+import { JwtAdapter } from '../controllers/auth/jwt.provider';
 
 @Module({
   imports: [
